% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_coder.R
\name{dummy_coder}
\alias{dummy_coder}
\title{Dummy Coder}
\usage{
dummy_coder(regexps, baseoff, levels, type = "factor", labels = FALSE)
}
\arguments{
\item{regexps}{A vector of strings containing the regular
expressions for each respective level one would like to
designate in the new dummy coded variable.}

\item{baseoff}{The vector to which the regular expressions
are applied to create the dummy coded variable.}

\item{levels}{The levels, each of which should correspond
to one of the 'regexps' provided, that will be included in
the new dummy coded variable.}

\item{type}{What variable type would you like the dummy
coded variable to be returned as? Options: factor,
integer, numeric, or character.}

\item{labels}{If 'type' is factor, optional labels
for the levels of the new dummy coded variable.}
}
\value{
'dummy_coder()' returns the dummy
coded variable as a vector, type determined
by 'type', that contains each of the corresponding values
based off 'baseoff'.
}
\description{
Dummy Coder
}
\examples{
exp_data <- datapreper::ex_ExperimentalData

exp_data$dummyone <- dummy_coder(regexps = c("^H", "^L"),
                            baseoff = exp_data$Cond,
                            levels = c(1, 0),
                            type = "factor",
                            labels = c("High", "Low"))

exp_data$dummytwo <- dummy_coder(regexps = c("W$", "L$"),
                            baseoff = exp_data$Cond,
                            levels = c(1, 0),
                            type = "numeric")


}
