% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_ExperimentalData}
\alias{ex_ExperimentalData}
\title{Example Experimental Data}
\format{
\subsection{\code{ex_ExperimentalData}}{

A data frame with 50 rows and 2 columns:
\describe{
\item{Cond}{column with four unique, two letter mappings for 4 conditions}
\item{value}{Randomly sampled values from 0 to 21}
}
}
}
\usage{
ex_ExperimentalData
}
\description{
Includes randomly sampled values and conditions, as if experimental data was put in long format.
}
\keyword{datasets}
