% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_QuestData}
\alias{ex_QuestData}
\title{Example Questionnaire Data}
\format{
\subsection{\code{ex_QuestData}}{

A data frame with 20 rows and 5 columns:
\describe{
\item{Q1, Q2, Q3, Q4, Q5}{Each column represents a question. Possible values were 1-5, randomly sampled, and NAs included}
}
}
}
\usage{
ex_QuestData
}
\description{
Sample questionnaire data, with the range of options being 1-5 and NAs included.
}
\keyword{datasets}
