% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_creator.R
\name{label_creator}
\alias{label_creator}
\title{Label Creator}
\usage{
label_creator(
  base,
  unique = NULL,
  order = 1:length(unique),
  uniquefront = FALSE,
  base_sep = " ",
  base_to_unique_sep = " ",
  copies = 1
)
}
\arguments{
\item{base}{Base(s) that are common to all labels.}

\item{unique}{Vector of unique elements for each unique label desired.}

\item{order}{Order in which to use the unique elements provided, by
default it applies unique elements in the order their provided.}

\item{uniquefront}{T/F, should unique element be put at the front
of the labels.}

\item{base_sep}{String used to separate base elements, if multiple
are provided.}

\item{base_to_unique_sep}{String use to separate base element(s) from
unique element.}

\item{copies}{Integer, how many copies of the labels should be returned.}
}
\value{
'label_creator()' returns a vector of labels (strings)
with a length equal to the length of 'unique'. If 'copies' is
greater than 1, then a list with the number of elements equal to
the number of copies requested will be returned, each element
being a vector according to the description provided above.
}
\description{
'label_creator()' allows one to quickly and personally create a
sequence of unique labels by merging a common base(s) with each
unique element from a vector.
}
\examples{
label_creator(base = "BFI",
              unique = c(1, 2, 3, 4, 5),
              uniquefront = FALSE,
              base_to_unique_sep = "_",
              copies = 2)

label_creator(base = c("PID", "Gran"),
              unique = c(1, 2, 3, 4, 5),
              order = c(5, 4, 3, 2, 1),
              uniquefront = FALSE,
              base_sep = "_",
              base_to_unique_sep = "#",
              copies = 1)

label_creator(base = "GYM",
              unique = c("North", "South", "East", "West"),
              uniquefront = TRUE,
              base_to_unique_sep = " ",
              copies = 1)


}
