% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_identifier.R
\name{series_identifier}
\alias{series_identifier}
\title{Series Identifier}
\usage{
series_identifier(vector, series, overlap = FALSE, indices = FALSE)
}
\arguments{
\item{vector}{Vector within which to identify the series.}

\item{series}{Vector of elements matching the series that
one wants to count or locate.}

\item{overlap}{T/F, can series within 'vector' overlap and still count?}

\item{indices}{T/F, do you want the indices of each identified series
to be returned?}
}
\value{
'series_identifier()' returns a list with 'total' equaling
the number of series identified within 'vector'; and if 'indices' is
True, then 'matchindices' is included in the list. 'matchindices' is itself
a list, with each element being one of the respective matches and its'
associated indices.
}
\description{
'series_identifier()' allows for the quick counting and location of
an overlapping or non-overlapping series within a vector of
elements.
}
\examples{
apply(datapreper::ex_QuestData, 1,
            series_identifier,
            series = c(4, 4),
            overlap = TRUE,
            indices = FALSE)

series_identifier(series = c(1, 3, 4),
                  vector = datapreper::ex_QuestData$Q5,
                  overlap = TRUE,
                  indices = TRUE)

}
